N-Template Release 10.0.0: Full Kotlin and EncryptedAssets Integration
We would like to thank Slim members Griffin Millender and Anthony King
in working with us to create the EncryptedAssets commits and would
offer to get on board with our theme template, allowing themers to
utilize the newest and greatest template from all of us.

To enable full Slim support, change build.gradle's line 14 from
false to true. It's as simple as that!

With release 10.0.0, we are pushing encrypted assets as a standard
because we have been receiving a bunch of complaints from people
regarding kangs and recompiled themes. Each of the theme's
asset/overlay folder is encrypted with an AES key that is registered
on compile time, so the themer nor the developers of Substratum will
be able to figure out the code on each compile.

This means that if you lose the source to your theme, you can no
longer extract it from an APK of yours. So please use BitBucket
or GitLab as a private repo to store your theme!

With this transition, we have also shifted to using the Kotlin
static programming language. This means that YOU CANNOT migrate from
an older version of the template, you MUST rebase. With this also
comes an update to Substratum to cherry-pick decryption keys, so
if you are using a modified template, you may get multiple dialogs
or theme instances - nothing if you use the official template.

Finally, we would like to remind you that this template will be for
Nougat onwards. Marshmallow support is now officially dropped.